+++ src/components/TheFooter.vue
@@ -15,7 +15,6 @@ onMounted(async () => {
 </script>
 
 <template>
-  <p>hej isabella</p>
   <div class="footerContainer">
     <img class="logoImage" src="../assets/MortensLogo.png" alt="">
     <ul>
